Refer the library paths to the linux user’s home folder
Being the PDK sample schematics in a "_prj" folder inside /home/<username>/.qucs (this is the default scenario), the simulator does not find the libraries because the include script in the schematics tells Qucs-S to go up two directories and look for the libraries.

Since the PDK library is installed in /home/<username>/.qucs/user_lib, Qucs-S won't find the PDK library. This problem is easily solved if the references are made to the home directory (i.e. simply replace "../.." with "~" in the example schematics and in the library).